feat(autosetup): fix subscription activation
Refs: CPLP-2344

+++ src/marketplace/Offers.Library/Service/OfferSetupService.cs
@@ -546,8 +546,11 @@ public class OfferSetupService : IOfferSetupService
             subscription.OfferSubscriptionStatusId = OfferSubscriptionStatusId.ACTIVE;
         });
 
-        offerSubscriptionRepository.RemoveOfferSubscriptionProcessData(offerSubscriptionId);
-
+        if (!offerDetails.InstanceData.IsSingleInstance)
+        {
+            offerSubscriptionRepository.RemoveOfferSubscriptionProcessData(offerSubscriptionId);
+        }
+        
         var notificationContent = JsonSerializer.Serialize(new
         {
             offerDetails.OfferId,
